requirements.txt
asn1crypto==0.24.0
attrs==18.2.0
Automat==0.7.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.4.2
cssselect==1.0.3
hyperlink==18.0.0
idna==2.5
incremental==17.5.0
lxml==4.3.0
numpy==1.15.4
pandas==0.23.4
parsel==1.5.1
pyasn1==0.4.5
pyasn1-modules==0.2.3
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pymongo==3.7.2
pyOpenSSL==18.0.0
python-dateutil==2.7.5
pytz==2018.9
queuelib==1.5.0
requests==2.16.5
requests-cache==0.4.13
scrapely==0.13.4
Scrapy==1.3.2
service-identity==18.1.0
six==1.12.0
Twisted==18.9.0
urllib3==1.21.1
w3lib==1.19.0
zope.interface==4.6.0
